+++ topic-operator/src/main/java/io/strimzi/operator/topic/v2/BatchingTopicController.java
@@ -62,6 +62,7 @@ public class BatchingTopicController {
 
     static final String FINALIZER = "strimzi.io/topic-operator";
     static final String MANAGED = "strimzi.io/managed";
+    static final String PAUSED = "strimzi.io/pause-reconciliation";
     static final String AUTO_CREATE_TOPICS_ENABLE = "auto.create.topics.enable";
     private final boolean useFinalizer;
 
@@ -110,6 +111,13 @@ public class BatchingTopicController {
                 || kt.getMetadata().getAnnotations() == null
                 || kt.getMetadata().getAnnotations().get(MANAGED) == null
                 || !"false".equals(kt.getMetadata().getAnnotations().get(MANAGED));
+    }
+
+    /* test */ static boolean isPaused(KafkaTopic kt) {
+        return kt.getMetadata() != null
+            && kt.getMetadata().getAnnotations() != null
+            && kt.getMetadata().getAnnotations().get(PAUSED) != null
+            && "true".equals(kt.getMetadata().getAnnotations().get(PAUSED));
     }
 
     private static boolean isForDeletion(KafkaTopic kt) {
@@ -418,12 +426,17 @@ public class BatchingTopicController {
 
         Map<ReconcilableTopic, Either<TopicOperatorException, Object>> results = new HashMap<>();
 
-        var partitionedByManaged = partitionedByDeletion.get(false).stream().collect(Collectors.partitioningBy(reconcilableTopic -> isManaged(reconcilableTopic.kt())));
+        var partitionedByManaged = partitionedByDeletion.get(false).stream()
+            .collect(Collectors.partitioningBy(reconcilableTopic -> isManaged(reconcilableTopic.kt())));
         var unmanaged = partitionedByManaged.get(false);
         addOrRemoveFinalizer(useFinalizer, unmanaged).forEach(rt -> putResult(results, rt, Either.ofRight(null)));
 
-        List<ReconcilableTopic> mayNeedUpdate = validateManagedTopics(partitionedByManaged);
+        // skip reconciliation of paused KafkaTopics
+        var partitionedByPaused = validateManagedTopics(partitionedByManaged).stream()
+            .collect(Collectors.partitioningBy(reconcilableTopic -> isPaused(reconcilableTopic.kt())));
+        partitionedByPaused.get(true).forEach(reconcilableTopic -> putResult(results, reconcilableTopic, Either.ofRight(null)));
 
+        var mayNeedUpdate = partitionedByPaused.get(false);
         var addedFinalizer = addOrRemoveFinalizer(useFinalizer, mayNeedUpdate);
 
         var currentStatesOrError = describeTopic(addedFinalizer);
@@ -961,25 +974,39 @@ public class BatchingTopicController {
             LOGGER.warnCr(reconcilableTopic.reconciliation(), "Updating status for unexpected exception", e);
             reason = e.getClass().getSimpleName();
         }
-        updateStatus(reconcilableTopic.reconciliation(), reconcilableTopic.kt(), reason, message);
+        Condition condition = new ConditionBuilder()
+            .withType("Ready")
+            .withStatus("False")
+            .withReason(reason)
+            .withMessage(message)
+            .withLastTransitionTime(StatusUtils.iso8601Now())
+            .build();
+        updateStatus(reconcilableTopic.reconciliation(), reconcilableTopic.kt(), condition);
     }
 
     private void updateStatusOk(ReconcilableTopic reconcilableTopic) {
-        updateStatus(reconcilableTopic.reconciliation(), reconcilableTopic.kt(), null, null);
+        Condition condition = new ConditionBuilder()
+            .withType(isPaused(reconcilableTopic.kt()) ? "ReconciliationPaused" : "Ready")
+            .withStatus("True")
+            .withLastTransitionTime(StatusUtils.iso8601Now())
+            .build();
+        updateStatus(reconcilableTopic.reconciliation(), reconcilableTopic.kt(), condition);
     }
 
     private void updateStatus(Reconciliation reconciliation,
                               KafkaTopic kt,
-                              String newReason,
-                              String newMessage) {
+                              Condition condition) {
         var oldStatus = kt.getStatus();
         Condition oldReadyCondition = oldStatus == null || oldStatus.getConditions() == null ? null : oldStatus.getConditions().stream().findFirst().orElse(null);
-        String newStatus = newReason != null || newMessage != null ? "False" : "True";
         if (oldStatus == null
                 || oldStatus.getObservedGeneration() != kt.getMetadata().getGeneration()
                 || oldStatus.getTopicName() == null
                 || oldReadyCondition == null
-                || isDifferentCondition(oldReadyCondition, newReason, newMessage, newStatus)) {
+                || isDifferentCondition(oldReadyCondition, condition)
+                || isPaused(kt)) {
+            long observedGeneration = oldStatus != null
+                ? !isPaused(kt) ? kt.getMetadata().getGeneration() : oldStatus.getObservedGeneration()
+                : !isPaused(kt) ? kt.getMetadata().getGeneration() : 0L;
             String newTopicName = !isManaged(kt) ? null
                     : oldStatus != null && oldStatus.getTopicName() != null ? oldStatus.getTopicName()
                     : topicName(kt);
@@ -988,15 +1015,9 @@ public class BatchingTopicController {
                         .withResourceVersion(null)
                     .endMetadata()
                     .editOrNewStatus()
-                        .withObservedGeneration(kt.getMetadata().getGeneration())
+                        .withObservedGeneration(observedGeneration)
                         .withTopicName(newTopicName)
-                        .withConditions(new ConditionBuilder()
-                                .withType("Ready")
-                                .withStatus(newStatus)
-                                .withReason(newReason)
-                                .withMessage(newMessage)
-                                .withLastTransitionTime(StatusUtils.iso8601Now())
-                                .build())
+                        .withConditions(condition)
                     .endStatus().build();
             LOGGER.debugCr(reconciliation, "Updating status with {}", updatedTopic.getStatus());
             long t0 = System.nanoTime();
@@ -1012,12 +1033,10 @@ public class BatchingTopicController {
     }
 
     private static boolean isDifferentCondition(Condition oldReadyCondition,
-                                                String newReason,
-                                                String newMessage,
-                                                String newStatus) {
+                                                Condition condition) {
         return !Objects.equals(oldReadyCondition.getType(), "Ready")
-                || !Objects.equals(oldReadyCondition.getStatus(), newStatus)
-                || !Objects.equals(oldReadyCondition.getReason(), newReason)
-                || !Objects.equals(oldReadyCondition.getMessage(), newMessage);
+                || !Objects.equals(oldReadyCondition.getStatus(), condition.getStatus())
+                || !Objects.equals(oldReadyCondition.getReason(), condition.getReason())
+                || !Objects.equals(oldReadyCondition.getMessage(), condition.getMessage());
     }
 }