+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Allow manual rolling of Kafka Connect and Kafka Mirror Maker 2 pods using the `strimzi.io/manual-rolling-update` annotation (supported only when `StableConnectIdentities` feature gate is enabled) 
 * Make sure brokers are empty before scaling them down
 * Update Cruise Control to 2.5.128
+* Add support for pausing `KafkaTopic` reconciliations with the UnidirectionalTopicOperator
 
 ### Changes, deprecations and removals